Add components AddRestaurant and Right_sideMenu Add pages MyRestaurants and CurrentOrders
